datas/backend/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.6
pytz==2024.1
sqlparse==0.4.4
